requirements.txt
numpy>=1.26.0
pandas>=2.2.0
PyYAML>=6.0.1
reportlab>=4.0.8
Pillow>=10.2.0
plotly>=5.22.0
streamlit>=1.36.0
matplotlib>=3.9.0
scipy>=1.13.0
statsmodels>=0.14.0
